+++ src/index.ts
@@ -1,2 +1,3 @@
 export { useRenderCount, useRemount } from './hooks/lifecycles';
 export { useModalState, useToggle } from './hooks/state';
+export { useLocalStorage } from './hooks/browser';